+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# Changelog
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial release.
+
+## 0.0.1+1
+
+* Legacy support to Dart SDK 2.12 & Flutter SDK <3.0
+
+## 0.1.0
+
+* Bumped the minimum required Dart SDK version to 2.17.